# public/api-docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Brainswarming API",
        "description": "API for the Brainswarming platform - a SaaS platform for teams to anonymously or openly submit, evaluate, and prioritize improvement ideas.",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/teams": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Get all teams for the authenticated user",
                "description": "Returns a list of all teams the authenticated user belongs to",
                "operationId": "getUserTeams",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "teams": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Team"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Create a new team",
                "description": "Creates a new team with the authenticated user as the founder",
                "operationId": "createTeam",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "team_code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Project Team"
                                    },
                                    "team_code": {
                                        "type": "string",
                                        "example": "NPT789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Team created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "team": {
                                            "$ref": "#/components/schemas/Team"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teams/join": {
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Join a team using a team code",
                "description": "Allows a user to join a team using a team code",
                "operationId": "joinTeamByCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "team_code"
                                ],
                                "properties": {
                                    "team_code": {
                                        "type": "string",
                                        "example": "DEV123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully joined the team",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully joined the team"
                                        },
                                        "team": {
                                            "$ref": "#/components/schemas/Team"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Team not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teams/join/{token}": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Join a team using an invite link token",
                "description": "Allows a user to join a team using an invite link token",
                "operationId": "joinTeamByLink",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Invite token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully joined the team",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully joined the team"
                                        },
                                        "team": {
                                            "$ref": "#/components/schemas/Team"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Team not found or invite token expired"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teams/invite/accept": {
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Accept a team invitation",
                "description": "Allows a user to accept a team invitation using an invite token",
                "operationId": "acceptTeamInvite",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invite_token"
                                ],
                                "properties": {
                                    "invite_token": {
                                        "type": "string",
                                        "example": "unique_invite_token_string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invitation accepted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invitation accepted successfully"
                                        },
                                        "team": {
                                            "$ref": "#/components/schemas/Team"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Team not found or invite token expired"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/teams/{teamId}/leave": {
            "delete": {
                "tags": [
                    "Teams"
                ],
                "summary": "Leave a team",
                "description": "Allows a user to leave a team they are a member of",
                "operationId": "leaveTeam",
                "parameters": [
                    {
                        "name": "teamId",
                        "in": "path",
                        "description": "ID of the team",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully left the team",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully left the team"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Cannot leave a team you founded"
                    },
                    "404": {
                        "description": "Team not found or user is not a member"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teams/{teamId}/invite/generate": {
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Generate a new invite link",
                "description": "Generates a new invite link for a team (requires admin privileges)",
                "operationId": "generateTeamInvite",
                "parameters": [
                    {
                        "name": "teamId",
                        "in": "path",
                        "description": "ID of the team",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "expires_in_days": {
                                        "description": "Number of days until the invite expires (default: 7)",
                                        "type": "integer",
                                        "example": 7
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invite link generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invite link generated successfully"
                                        },
                                        "invite_token": {
                                            "type": "string",
                                            "example": "unique_invite_token_string"
                                        },
                                        "invite_link": {
                                            "type": "string",
                                            "example": "https://example.com/teams/join/unique_invite_token_string"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-06-17T12:00:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - User is not an admin"
                    },
                    "404": {
                        "description": "Team not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}